fix: split local multipart operations
